feat(app): 金额统一为整数分（Room v13，amount_minor INTEGER）
存储层（Room v12 → v13）
- Bill.amount / Budget.amount / BillTemplate.amount / Account.balance 由 Double 改为
  amountMinor / balanceMinor（Long，列名 amount_minor / balance_minor）
- MIGRATION_12_13 按官方建议重建四张表（SQLite 不能改列类型，且 Room 校验列亲和性）。
  顺序上先把 bills 数据搬入无约束中转表再重建 accounts，避免外键开启时 DROP 被引用表失败；
  回填统一 CAST(ROUND(amount*100) AS INTEGER)；@Entity 声明的全部索引手动重建
- 迁移 DDL 已与 KSP 生成的 schemas/13.json 逐列比对一致；新增 schemas/13.json
- 新增 util/Money：parseMinor（元字符串→分，正则拆分后整数运算，禁 toDouble()*100）、
  yuanToMinor（HALF_UP，仅用于语音等已是 Double 的边界）、format / formatPlain /
  toYuanInputString（纯整数拆分 + 手工千分位，不依赖 Locale）

数据与网络层
- BillDao 全部 SUM 改对 amount_minor，返回 Long；DailyCategoryAmount.total 改 Long
- Bill/Account/Budget 仓库接口与实现金额类型改 Long 分；余额联动（nudgeAccount）整数运算
- DTO 双字段：BillDTO/AccountDTO/TemplateDTO/BudgetDTO 增加 amountMinor/balanceMinor（权威），
  旧 amount/balance（元）保留用于旧服务端回退，经 resolvedXxxMinor 统一解析；
  CreateBillRequest/UpdateAccountRequest/UpsertBudgetRequest 只发分
- SyncManager 上行发分、下行优先读分；洞察 DTO（MonthlyReview/Spike/SingleBill）增加 Minor 字段

ViewModel / UI
- Bookkeeping/Budget/QuickAdd/Assets/Ai 各 VM 的 totals、预算比较、余额联动全部整数化；
  进度百分比在最后一步转 Double 相除，避免整数相除截断
- 输入解析统一 Money.parseMinor（键盘/语音/模板回填）；展示统一 Money.format /
  formatPlain，全库不再有 "¥" + %.2f 的浮点格式化
- 顺带修复 4 处潜在运行时崩溃：String.format("%.2f", Long) 在运行时会抛
  IllegalFormatConversionException（MonthChartPager / PlanScreen / SummaryBar 等）

测试
- 新增 MoneyTest（9 项：漂移样本、非法输入、往返无损、动画取整）
- 既有 VM/仓库/同步测试全部改为分语义（工厂函数入参保持「元」、边界处换算）
- :app:compileDebugKotlin / :app:compileDebugUnitTestKotlin 通过
- :app:testDebugUnitTest 109 项，失败 6 项与改造前基线完全一致（同名同断言，既有问题）

+++ app/src/main/java/com/example/rinklnote/data/db/entity/DailyCategoryAmount.kt
@@ -6,5 +6,5 @@ data class DailyCategoryAmount(
     @ColumnInfo(name = "category_name")
     val categoryName: String,
     @ColumnInfo(name = "total")
-    val total: Double
+    val total: Long
 )
